# Switch to transforming on a per character basis for non-valid utf8 messages