This PR adjusts the work done in #21. An edge case can happen with blanket coercing a non-valid utf8 string into a valid utf8 string as mentioned here : https://www.erlang.org/doc/apps/stdlib/unicode_usage.html#double-utf-8-encoding

See the test case added for an example. Unfortunately, this means when the string in question is not valid utf8 the safest way to perform the conversion is char by char. This doesn't seem to have a huge performance penalty.