@DarkMarmot 2023-03-16
> This PR adjusts the work done in #21. An edge case can happen with blanket coercing a non-valid utf8 string into a valid utf8 string as mentioned here : https://www.erlang.org/doc/apps/stdlib/unicode_usage.html#double-utf-8-encoding
> 
> See the test case added for an example. Unfortunately, this means when the string in question is not valid utf8 the safest way to perform the conversion is char by char. This doesn't seem to have a huge performance penalty.

The nice thing about running `String.valid?` on the whole puppy is that it lets us generate quick `InvalidMessage` for pure junk I think? 

I think we might want to keep the latin1 naming as well. If i understand correctly, we can generate incorrect characters for latin9, latin15, etc if we encounter those encodings?

And running benchmarks, this looks like it would drop our performance by at least 2x. I get 55k ips down to 26k ips on my laptop when I toggle the setting. I don't know if there's a good answer here, but it is a pretty heavy hit for us :(