@starbelly 2023-03-16
> The nice thing about running `String.valid?` on the whole puppy is that it lets us generate quick `InvalidMessage` for pure junk I think?

Oh you mean in the case the string is hopeless. Maybe. 🤔 

> I think we might want to keep the latin1 naming as well. If i understand correctly, we can generate incorrect characters for latin9, latin15, etc if we encounter those encodings?

That makes sense. I'm happy to change it back, I figured it made less sense after the change I made.

> 
> And running benchmarks, this looks like it would drop our performance by at least 2x. I get 55k ips down to 26k ips on my laptop when I toggle the setting. I don't know if there's a good answer here, but it is a pretty heavy hit for us :(

huh, so here's what I get : 

## Before

```
Name                 ips        average  deviation         median         99th %
raw             164.96 K        6.06 μs   ±130.77%        5.83 μs        6.79 μs
new              84.25 K       11.87 μs    ±20.85%       11.25 μs       20.42 μs
new-copy         78.55 K       12.73 μs    ±19.33%       12.13 μs       19.92 μs
round-trip       40.52 K       24.68 μs    ±14.16%       23.50 μs       37.35 μs
new-alt          24.13 K       41.44 μs     ±6.74%       40.67 μs       46.33 μs

Comparison:
raw             164.96 K
new              84.25 K - 1.96x slower +5.81 μs
new-copy         78.55 K - 2.10x slower +6.67 μs
round-trip       40.52 K - 4.07x slower +18.62 μs
new-alt          24.13 K - 6.84x slower +35.37 μs
```

## After

```
Name                 ips        average  deviation         median         99th %
raw             162.96 K        6.14 μs   ±112.50%        5.88 μs        8.21 μs
new              87.54 K       11.42 μs    ±49.08%       10.25 μs       43.25 μs
new-copy         78.99 K       12.66 μs    ±18.22%       12.08 μs       18.75 μs
round-trip       40.49 K       24.70 μs    ±15.44%       23.25 μs       37.54 μs
new-alt          24.36 K       41.05 μs     ±6.00%       40.21 μs       50.75 μs

Comparison:
raw             162.96 K
new              87.54 K - 1.86x slower +5.29 μs
new-copy         78.99 K - 2.06x slower +6.52 μs
round-trip       40.49 K - 4.03x slower +18.56 μs
new-alt          24.36 K - 6.69x slower +34.92 μs
```

Maybe try running the benchmark again? I've odd results with Benchee (i.e., yesterday with transcode functions)...