@starbelly 2023-03-16
> > The nice thing about running `String.valid?` on the whole puppy is that it lets us generate quick `InvalidMessage` for pure junk I think?
> 
> Oh you mean in the case the string is hopeless. Maybe. 🤔

@DarkMarmot What we could do is bail early in the transcode function if we find a non-utf8 character. The main thing was avoiding potentially iterating twice (especially since transcode does exactly what String.valid?/1 does, accumulation and transcoding aside that is).