@camilleryr 2023-03-16
Benchmarks 

on main
```
Name                 ips        average  deviation         median         99th %
raw              82.33 K       12.15 μs    ±57.62%       10.99 μs       40.99 μs
new              51.71 K       19.34 μs    ±39.79%       16.99 μs       54.99 μs
new-copy         44.74 K       22.35 μs    ±41.68%       18.99 μs       59.99 μs
round-trip       20.87 K       47.91 μs    ±33.31%       44.99 μs      122.99 μs
new-alt          14.42 K       69.34 μs    ±29.13%       62.99 μs      164.99 μs
```

feature branch
```
Name                 ips        average  deviation         median         99th %
raw              75.82 K       13.19 μs    ±70.39%          11 μs          51 μs
new-copy         43.00 K       23.26 μs    ±50.99%          19 μs          70 μs
new              19.18 K       52.14 μs    ±30.97%          48 μs         120 μs
round-trip       17.65 K       56.67 μs    ±33.95%          50 μs         139 μs
new-alt          12.58 K       79.46 μs    ±30.40%          70 μs         178 μs
```