@starbelly 2023-03-22
After a lot of thought we have decided to not support this as an option, instead users of this lib should perform encoding conversion ahead of passing a payload into HL7. We will keep the validation, but open up another PR to remove the accept_latin1 option. 

Cheers.